+++ stepfunctions/src/main/scala/awscala/stepfunctions/Activity.scala
@@ -4,7 +4,7 @@ import awscala.stepfunctions.ArnFormat.ResourceArn
 import com.amazonaws.services.stepfunctions.model.DeleteActivityRequest
 
 case class Activity(arn: String) {
-  def name: String = ArnFormat.parseArn(arn, ResourceArn)
+  val name: String = ArnFormat.parseArn(arn, ResourceArn)
 
   def delete()(implicit steps: StepFunctions): Unit =
     steps.deleteActivity(new DeleteActivityRequest().withActivityArn(arn))
