+++ stepfunctions/src/main/scala/awscala/stepfunctions/StateMachine.scala
@@ -1,13 +1,15 @@
 package awscala.stepfunctions
 
-import awscala.DateTime
+import java.util
+
+import awscala.{ DateTime, Sequencer }
 import awscala.stepfunctions.ArnFormat.ResourceArn
 import com.amazonaws.services.stepfunctions.model._
 
 import scala.collection.JavaConverters._
 
 case class StateMachine(arn: String) {
-  def name: String = ArnFormat.parseArn(arn, ResourceArn)
+  val name: String = ArnFormat.parseArn(arn, ResourceArn)
 
   def startExecution(input: String, name: String = null)(implicit steps: StepFunctions): Execution = {
     val exec =
@@ -22,12 +24,19 @@ case class StateMachine(arn: String) {
 
   def execution(name: String)(implicit steps: StepFunctions): Option[Execution] = executions().find(_.name == name)
 
-  def executions()(implicit steps: StepFunctions): Seq[Execution] =
-    steps
-      .listExecutions(new ListExecutionsRequest().withStateMachineArn(arn))
-      .getExecutions
-      .asScala
-      .map(e => Execution(e.getExecutionArn, new DateTime(e.getStartDate)))
+  def executions()(implicit steps: StepFunctions): Seq[Execution] = {
+    object ExecutionsSequencer extends Sequencer[ExecutionListItem, ListExecutionsResult, String] {
+      val base = new ListExecutionsRequest().withStateMachineArn(arn)
+      def getInitial: ListExecutionsResult = steps.listExecutions(base)
+
+      def getMarker(r: ListExecutionsResult): String = r.getNextToken
+
+      def getFromMarker(marker: String): ListExecutionsResult = steps.listExecutions(base.withNextToken(marker))
+
+      def getList(r: ListExecutionsResult): util.List[ExecutionListItem] = r.getExecutions
+    }
+    ExecutionsSequencer.sequence.map(e => Execution(e.getExecutionArn, new DateTime(e.getStartDate)))
+  }
 
   def definition()(implicit steps: StepFunctions): String =
     steps.describeStateMachine(new DescribeStateMachineRequest().withStateMachineArn(arn)).getDefinition
